creamos el metodo reistroUsuario y algunas clases e interfaces más de PlatoAlimento en controller, service...

+++ src/main/java/com/nerea/mydish/repository/entity/AlimentoEntity.java
@@ -13,5 +13,5 @@ public class AlimentoEntity {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long idAlimento;
 	private String nombre;
-	private Long cal;
+	private Double cal;
 }
